2024 07 01 doc constants

+++ test/src/LibDecimalFloat.decimal.t.sol
@@ -1,14 +1,14 @@
 // SPDX-License-Identifier: CAL
 pragma solidity =0.8.25;
 
-import {LibDecimalFloat, COEFFICIENT_MASK, ExponentOverflow, NORMALIZED_MAX} from "src/lib/LibDecimalFloat.sol";
+import {LibDecimalFloat, ExponentOverflow, NORMALIZED_MAX} from "src/lib/LibDecimalFloat.sol";
 
-import {Test, console} from "forge-std/Test.sol";
+import {Test} from "forge-std/Test.sol";
 
 contract LibDecimalFloatDecimalTest is Test {
     /// Round trip from/to decimal values without precision loss
     function testFixedDecimalRoundTripLossless(uint256 value, uint8 decimals) external pure {
-        value = bound(value, 0, NORMALIZED_MAX);
+        value = bound(value, 0, uint256(NORMALIZED_MAX));
 
         (int256 signedCoefficient, int256 exponent, bool lossless0) =
             LibDecimalFloat.fromFixedDecimalLossy(value, decimals);
